requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
python-multipart==0.0.16
pillow==10.4.0
numpy==1.24.4
opencv-python-headless==4.10.0.84
